requirements.txt
peewee
python-dotenv
psycopg2-binary